# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LIS",
        "description": "Laboratory Information Service API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8089",
            "description": "Local server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/api/v1/lis-service/tests/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get a test by id",
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Update a test",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete a test",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/lis-service/laboratory-tests/{id}": {
            "get": {
                "tags": [
                    "Laboratory tests"
                ],
                "summary": "Get a laboratory-test offering by id",
                "operationId": "getById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaboratoryTestResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Laboratory tests"
                ],
                "summary": "Update a laboratory-test offering",
                "operationId": "update_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLaboratoryTest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaboratoryTestResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Laboratory tests"
                ],
                "summary": "Delete a laboratory-test offering",
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/lis-service/laboratories/{id}": {
            "get": {
                "tags": [
                    "Laboratories"
                ],
                "summary": "Get a laboratory by id",
                "operationId": "getById_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaboratoryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Laboratories"
                ],
                "summary": "Update a laboratory",
                "operationId": "update_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLaboratory"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaboratoryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Laboratories"
                ],
                "summary": "Delete a laboratory",
                "operationId": "delete_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/v1/lis-service/laboratory-tests": {
            "get": {
                "tags": [
                    "Laboratory tests"
                ],
                "summary": "List laboratory-test offerings (paged), optionally filtered",
                "operationId": "list_1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "laboratoryId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "testId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseLaboratoryTestResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Laboratory tests"
                ],
                "summary": "Register a test as offered by a laboratory",
                "operationId": "create_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLaboratoryTest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaboratoryTestResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/laboratories": {
            "get": {
                "tags": [
                    "Laboratories"
                ],
                "summary": "List laboratories (paged), optionally filtered by district, type and active flag",
                "operationId": "list_2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "districtId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "labTypeId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseLaboratoryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Laboratories"
                ],
                "summary": "Create a laboratory",
                "operationId": "create_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLaboratory"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LaboratoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UpdateTest": {
                "type": "object",
                "description": "Update a test",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 256,
                        "minLength": 0
                    },
                    "abbreviation": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "shortTitle": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "isCompositeTest": {
                        "type": "boolean"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "TestResponse": {
                "type": "object",
                "description": "Test",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "loincCode": {
                        "type": "string"
                    },
                    "abbreviation": {
                        "type": "string"
                    },
                    "shortTitle": {
                        "type": "string"
                    },
                    "isCompositeTest": {
                        "type": "boolean"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UpdateLaboratoryTest": {
                "type": "object",
                "description": "Update a laboratory-test offering",
                "properties": {
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "LaboratoryTestResponse": {
                "type": "object",
                "description": "Laboratory-test offering",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "laboratoryId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "testId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UpdateLaboratory": {
                "type": "object",
                "description": "Update a laboratory",
                "properties": {
                    "labName": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "districtId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "comment": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "labTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "LaboratoryResponse": {
                "type": "object",
                "description": "Laboratory",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "labCode": {
                        "type": "string"
                    },
                    "labName": {
                        "type": "string"
                    },
                    "districtId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "labTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "CreateLaboratoryTest": {
                "type": "object",
                "description": "Register a test as offered by a laboratory",
                "properties": {
                    "laboratoryId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "testId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "laboratoryId",
                    "testId"
                ]
            },
            "CreateLaboratory": {
                "type": "object",
                "description": "Create a laboratory",
                "properties": {
                    "labCode": {
                        "type": "string",
                        "maxLength": 5,
                        "minLength": 0
                    },
                    "labName": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "districtId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "comment": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "labTypeId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "districtId",
                    "labCode",
                    "labName"
                ]
            },
...
            "PagedResponseLaboratoryTestResponse": {
                "type": "object",
                "description": "Paged response",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "List of items",
                        "example": "[{\"id\":1,\"name\":\"Item 1\"}, ...]",
                        "items": {
                            "$ref": "#/components/schemas/LaboratoryTestResponse"
                        }
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Total number of items",
                        "example": 100
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Total number of pages",
                        "example": 10
                    },
                    "currentPage": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Current page number",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of items per page",
                        "example": 20
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                }
            },
            "PagedResponseLaboratoryResponse": {
                "type": "object",
                "description": "Paged response",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "List of items",
                        "example": "[{\"id\":1,\"name\":\"Item 1\"}, ...]",
                        "items": {
                            "$ref": "#/components/schemas/LaboratoryResponse"
                        }
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Total number of items",
                        "example": 100
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Total number of pages",
                        "example": 10
                    },
                    "currentPage": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Current page number",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of items per page",
                        "example": 20
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}